Zip/requirements.txt
asgiref==3.5.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.1
django-admin-interface==0.18.4
django-colorfield==0.6.3
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-multi-email-field==0.6.2
django-session-timeout==0.1.0
django-widget-tweaks==1.4.12
dnspython==2.2.1
email-validator==1.2.1
idna==3.3
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-decouple==3.6
pytz==2022.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
wfastcgi==3.0.0
whitenoise==5.3.0
xlwt==1.3.0
